maj voisinnage

+++ src/main/java/com/mmeg/glyphes/optimizer/pojo/Solution.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class Solution {
 	private Glyphage glyphage;
 	private Mob mob;
-	private long valeurFonctionObjectif;
+	private double valeurFonctionObjectif;
 
 	private long pvTotaux;
 	private long defTotaux;
